Version alpha-2 release
Note:  This pre-release version of Slide Set should be considered
unstable, and may not be suitable for critical work. In particular, ROI
set files created using this version may not be compatible with future
versions of Slide Set or ImageJ. This plugin works only with ImageJ
2.0.0-beta-7 (available from http://developer.imagej.net/downloads). It is
not compatible with later releases, including beta-7 patches (such as
beta-7.5). It is not compatible with ImageJ2 run through Fiji.

Updates from the previous version:

- Manders' Colocalization Coefficients command
- Improved correlation algorithm
- Documentation pages
- Various stability and usability improvements