Version 1.4.2 release
1.4.2 Release Notes
~~~~~~~~~~~~~~~~~~~

SlideSet is now built with Java 8 for compatibility with the latest
versions of ImageJ and ImageJ2. This version is will not run with older
ImageJ versions built with Java 6. To check which version of Java is
running, select `Help > About ImageJ` from the menu. If the Java version
number is `1.6.X_XX`, download a fresh installation of ImageJ to upgrade
to a Java 8 version.

Updates from the previous version:

- Now built with Java 8
- Compatibility fixes for ImageJ2 updates